requirements.txt
flask
dash
plotly
psutil
speedtest-cli
